Add Hello World at "/"
This change is the initial change made to the application. The controller is available and working fine. We can then start working our way up and close / merge this feature in develop branch.
This change has been made using TDD, which will be the case for all the project.
We chose to only check the string value of response message since this will be removed in the future.